requirements.txt
blis==0.7.4
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.5
numpy==1.21.2
packaging==21.0
pathy==0.6.0
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
python-dotenv==0.19.1
regex==2021.10.8
requests==2.26.0
scipy==1.7.1
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
textblob==0.15.3
thinc==8.0.10
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2
